Platform: Windows: ServiceCore: cast invalid PID

+++ plugins/platform/windows/WindowsServiceCore.cpp
@@ -38,7 +38,7 @@ static DWORD findWinlogonProcessId( DWORD sessionId )
 {
 	PWTS_PROCESS_INFO processInfo = nullptr;
 	DWORD processCount = 0;
-	DWORD pid = -1;
+	DWORD pid = static_cast<DWORD>( -1 );
 
 	if( WTSEnumerateProcesses( WTS_CURRENT_SERVER_HANDLE, 0, 1, &processInfo, &processCount ) == false )
 	{
